RleDataFrame now contains DFrame instead of DataFrame
The RleDataFrame class now indirectly extends DataFrame (via the new
DFrame class) instead of being a direct DataFrame extension. This is
in preparation for future changes to the DataFrame class hierarchy (see
commit 36837bdf to S4Vectors for more information).

This change is transparent to the end user and is not expected to break
any existing code or serialized object.